[fio fromtree] dockerapp: Fix docker-app invocation
The version of docker-app I built and tested this with was done
incorrectly. docker-app includes two binaries:

 https://github.com/docker/app/tree/master/cmd

There's a "standalone" and "docker-app". I mistakenly built the
"docker-app" which is intended to become a plugin to the docker
command. However, that's still marked "experimental". So, for now,
we should use the standalone version. That version makes the invocations
look like `docker-app render` rather than the current
`docker-app app render`.

In the near future we should change this to handle either approach.

Signed-off-by: Andy Doan <andy@foundries.io>
Signed-off-by: Ricardo Salveti <ricardo@foundries.io>