[fio extras] aktualizr-lite: daemon: skip older known images (rollback)
Check if the target image is known when running in daemon mode and skip
the update process if so. This is done as a way to detect rollbacks
after applying an update as the image will be stored locally in the
installation log table.

The correct implementation requires api/storage changes in order to
support broken/bad states in the installation log table, which needs to
be done in upstream first (and then drop/update this patch).

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>